+++ src/FORMAT_TYPE.mjs
@@ -1,7 +1,0 @@
-export const FORMAT_TYPE_COLOR = "color";
-
-export const FORMAT_TYPE_EMAIL = "email";
-
-export const FORMAT_TYPE_TEXT = "text";
-
-export const FORMAT_TYPE_URL = "url";