+++ CHANGELOG.md
@@ -6,6 +6,14 @@ Changes:
 
 \-
 
+## v2023-05-30-1
+
+Changes:
+
+- Rename to `flux-value-format`
+- Url label/title
+- Natvie `adoptedStyleSheets`
+
 ## v2023-05-25-1
 
 Changes: